requirement.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
astroid==1.5.3
autopep8==1.3.2
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
basho-erlastic==2.1.1
billiard==3.3.0.23
bleach==2.0.0
celery==3.1.26.post2
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
contextlib2==0.5.5
coreapi==2.3.1
coreschema==0.0.4
decorator==4.1.2
Django==1.11.4
django-ajax-selects==1.6.0
django-allauth==0.32.0
django-celery==3.2.1
django-cors-middleware==1.3.1
django-filter==1.0.4
django-memoize==2.1.0
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
django-silk==1.0.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
dodgy==0.1.9
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
html5lib==0.999999999
httpie==0.9.9
idna==2.6
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
isort==4.2.15
itypes==1.1.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
-e git://github.com/jupyter/jupyter-drive.git@545813377cb901235e8ea81f83b0ac7755dbd7a9#egg=jupyterdrive
kombu==3.0.37
lazy-object-proxy==1.3.1
Markdown==2.6.8
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.7.4
MySQL-python==1.2.5
nbconvert==5.3.1
nbformat==4.4.0
oauthlib==2.0.2
openapi-codec==1.3.2
pandocfilters==1.4.2
pathlib2==2.3.0
pep8-naming==0.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
prospector==0.12.7
psycopg2==2.7.3
ptyprocess==0.5.2
pycodestyle==2.4.0
pycountry==17.5.14
pydocstyle==2.0.0
PyExecJS==1.4.0
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.5.2
pylint==1.7.2
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==0.7.2
pylint-flask==0.5
pylint-plugin-utils==0.2.6
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
raven==6.1.0
requests==2.18.4
requests-oauthlib==0.8.0
requirements-detector==0.5.2
scandir==1.5
setoptconf==0.2.0
simplegeneric==0.8.1
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sqlparse==0.2.3
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.15
vine==1.1.4
virtualenv==15.0.2
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.10.11
